EIP1-3666 Refactored abbreviation temp cert to full temporary certificate
