+++ src/main/kotlin/uk/gov/dluhc/printapi/service/temporarycertificate/TemplateDetails.kt
@@ -1,4 +1,4 @@
-package uk.gov.dluhc.printapi.service.tempcert
+package uk.gov.dluhc.printapi.service.temporarycertificate
 
 data class TemplateDetails(
     val path: String,
