+++ src/main/resources/application.yml
@@ -49,7 +49,7 @@ jobs:
     name: "ProcessPrintResponses"
     cron: "0 15/30 * * * *" # every 30 minutes starting at 15 minutes past the hour - see analysis and recommendations in EIP1-2515
 
-temp-cert:
+temporary-certificate:
   explainer-pdf:
     english:
       path: "templates/Explainer Document (English).pdf"
